third_party/blink/web_tests/http/tests/devtools/geolocation-emulation-tests-expected.txt

Tests that geolocation emulation with latitude and longitude works as expected.


Running: testGeolocationUnavailable
Position unavailable

Running: testOverridenGeolocation
Latitude: 50 Longitude: 100

Running: testInvalidParam
error: Protocol Error: Invalid type of argument 'latitude' for method 'Emulation.setGeolocationOverride' call. It must be 'number' but it is 'boolean'.

Running: testInvalidGeolocation
Latitude: 50 Longitude: 100

Running: testNoOverride
Override was cleared correctly.
